feat: add validation to form
