+++ src/components/order/order.ts
@@ -1,13 +1,18 @@
 class orderWindow {
+  private container: HTMLElement;
+
+  constructor () {
+    this.container = <HTMLElement>document.querySelector('.main');
+  }
   openWindow() {
-    const main = <HTMLElement>document.querySelector('.main');
+    //const main = <HTMLElement>document.querySelector('.main');
     const modalWindow = document.createElement('div');
     modalWindow.classList.add('modal');
 
     const titleModal = document.createElement('h3');
     titleModal.classList.add('modal__title');
     titleModal.innerText = 'Personal details';
-    modalWindow.append(titleModal);
+    modalWindow.append(titleModal);    
 
     const form = document.createElement('form');
 
@@ -15,12 +20,11 @@ class orderWindow {
     inputName.classList.add('input-name', 'input');
     inputName.placeholder = 'Name';
     inputName.type = 'text';
-    //inputName.pattern = '[A-Za-zА-Яа-яЁё]{3,} [A-Za-zА-Яа-яЁё]{3,}.*';
 
     const inputPhone = document.createElement('input');
     inputPhone.classList.add('input-phone', 'input');
     inputPhone.placeholder = 'Phone number';
-    inputPhone.type = 'number';
+    inputPhone.type = 'text';
 
     const inputAdress = document.createElement('input');
     inputAdress.classList.add('input-adress', 'input');
@@ -31,13 +35,9 @@ class orderWindow {
     inputMail.placeholder = 'E-mail';
     inputMail.type = 'email';
 
-    form.append(inputName, inputPhone, inputAdress, inputMail);
-    modalWindow.append(form);
-
     const creditTitle = document.createElement('h3');
     creditTitle.classList.add('modal__credit');
     creditTitle.innerText = 'Credit card details';
-    modalWindow.append(creditTitle);
 
     const creditCard = document.createElement('div');
     creditCard.classList.add('credit-card');
@@ -47,9 +47,10 @@ class orderWindow {
 
     const cardIcon = document.createElement('img');
     cardIcon.classList.add('card-icon');
+    cardIcon.src = 'https://play-lh.googleusercontent.com/baXy546Srucl3vM1yaHr060eBL9_mrk0NH2GGRRCMTrKbekbx2pI77WCaXmNwUqnqQ';
 
     const numberField = document.createElement('input');
-    numberField.classList.add('number-field');
+    numberField.classList.add('number-field', 'input');
     numberField.placeholder = 'Card number';
     numberField.type = 'number';
 
@@ -59,74 +60,201 @@ class orderWindow {
     validBlock.classList.add('valid-block');
 
     const dateInput = document.createElement('input');
-    dateInput.classList.add('date-input');
+    dateInput.classList.add('date-input', 'input');
     dateInput.placeholder = 'Validaty';
-    //dateInput.type = 'number';
 
     const cvvInput = document.createElement('input');
-    cvvInput.classList.add('cvv-input');
+    cvvInput.classList.add('cvv-input', 'input');
     cvvInput.placeholder = 'CVV';
     cvvInput.type = 'number';
     validBlock.append(dateInput, cvvInput);
 
     creditCard.append(cardNumber, validBlock);
-    modalWindow.append(creditCard);
 
     const confirmBtn = document.createElement('button');
     confirmBtn.classList.add('confirm-button');
     confirmBtn.innerText = 'Confirm';
+    confirmBtn.type = 'submit';
 
-    modalWindow.append(confirmBtn);
+    form.append(inputName, inputPhone, inputAdress, inputMail, creditTitle, creditCard, confirmBtn);
+    modalWindow.append(form);   
+    
+    const modalBack = document.createElement('div');
+    modalBack.classList.add('modal-back', 'modal-back-open');
 
-    main.append(modalWindow);
+    this.container.append(modalWindow, modalBack);
 
     //validation
-    // inputName.addEventListener('blur', function() {
-    //   const arr = inputName.value.split(' ');
-    //   if (arr.length > 1) {
-    //     for (let i = 0; i < arr.length; i++) {
-    //       if (arr[i].length > 2) {
-    //         console.log('yes');
-    //       }
-    //     } 
-    //   } else {
-    //     console.log('no!')
-    //   }
-           
-    // })
 
-    inputPhone.addEventListener('blur', function() {
-      if (inputPhone.value[0] !== '+' || inputPhone.value.length < 10) {
-        console.log(typeof(inputPhone.value));
-        console.log('no');
+    const formInputs = document.querySelectorAll('.input');
+
+    //name
+    function checkName (name: string) {
+      const array = name.split(' ');
+      if (array.length >= 2) {
+        const filterArr = array.filter(el => el.length >= 3);
+        if (array.length === filterArr.length) {
+          return true;
+        }
       } else {
-        console.log('yes');
+        return false;
       }
-    })
+    }
+
+    //phone
+    function checkPhone (phone: string) {
+      const PHONE_REGEXP = /^(\+)((\d{2,3}) ?\d|\d)(([ -]?\d)|( ?(\d{2,3}) ?)){7,15}\d$/iu;
+      return PHONE_REGEXP.test(phone);
+    }
+
+    //adress
+    function checkAdress (adress: string) {
+      const array = adress.split(' ');
+      if (array.length >= 3) {
+        const filterArr = array.filter(el => el.length >= 5);
+        if (array.length === filterArr.length) {
+          return true;
+        }
+      } else {
+        return false;
+      }
+    }
 
     //mail
-    const EMAIL_REGEXP = /^(([^<>()[\].,;:\s@"]+(\.[^<>()[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/iu;
-    function validEmail(value: string) {
-      return EMAIL_REGEXP.test(value);
+    function checkMail (mail: string) {
+      const EMAIL_REGEXP = /^(([^<>()[\].,;:\s@"]+(\.[^<>()[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/iu;
+      return EMAIL_REGEXP.test((mail).toLowerCase());
     }
-    function checkMail() {
-      if (validEmail(inputMail.value)) inputMail.style.borderColor = '#46760A';
-      else inputMail.style.borderColor = 'red';
+
+    function createErrorText (field: string, container: HTMLElement) {  
+      const err = <HTMLElement>document.querySelector('.error-text');    
+      if (err) {
+        return;
+      } else {
+        const errorText = document.createElement('span');
+        errorText.classList.add('error-text');        
+        errorText.innerText = `Enter a correct ${field}`;      
+        container.after(errorText);
+      }      
     }
-    inputMail.addEventListener('input', checkMail);
+
+    function deleteErrorText () {
+      const err = <HTMLElement>document.querySelector('.error-text');
+      if (err) {
+        err.remove();
+      }
+    }
+
+    form.onsubmit = function () {
+      const nameValue = inputName.value;
+      const mailValue = inputMail.value;
+      const phoneValue = inputPhone.value;
+      const adressValue = inputAdress.value;
+
+      const emptyValue = Array.from(formInputs).filter((input: any) => input.value === '');
+      
+      formInputs.forEach(function (input: any) {
+        if (input.value === '') {
+          input.classList.add('error');
+        } else {
+          input.classList.remove('error');
+        }
+      })
+
+      if (emptyValue.length !== 0) {
+        return false;
+      }
+
+      if (!checkName(nameValue)) {
+        inputName.classList.add('error');
+        createErrorText('name', inputName);
+        return false;
+      } else {
+        inputName.classList.remove('error');
+        deleteErrorText();        
+      }
+
+      if (!checkPhone(phoneValue)) {
+        inputPhone.classList.add('error');
+        createErrorText('phone', inputPhone);
+        return false;
+      } else {
+        inputPhone.classList.remove('error');       
+        deleteErrorText();        
+      }      
+
+      if (!checkAdress(adressValue)) {
+        inputAdress.classList.add('error');
+        createErrorText('adress', inputAdress);
+        return false;
+      } else {
+        inputAdress.classList.remove('error');
+        deleteErrorText();
+      }
+
+      if (!checkMail(mailValue)) {
+        inputMail.classList.add('error');        
+        createErrorText('e-mail', inputMail);
+        return false;
+      } else {
+        inputMail.classList.remove('error');
+        deleteErrorText();
+      }
+
+      if (numberField.value.length !== 16) {
+        numberField.classList.add('error');        
+        createErrorText('number', cardNumber);
+        return false;
+      } else {
+        numberField.classList.remove('error');        
+        deleteErrorText();
+      }
+
+      if (Number(dateInput.value.split('/')[0]) > 12 || dateInput.value.length < 5) {
+        dateInput.classList.add('error');       
+        createErrorText('date', validBlock);
+        return false;
+      } else {
+        dateInput.classList.remove('error');
+        deleteErrorText();
+      }
+
+      if (cvvInput.value.length !== 3) {
+        cvvInput.classList.add('error');       
+        createErrorText('cvv', validBlock);
+        return false;
+      } else {
+        cvvInput.classList.remove('error');
+        deleteErrorText();
+      }
+      
+      modalWindow.innerHTML = 'Thank you for the order. Redirect to home page...';
+
+    }    
     
     //cardNUmber
+
+    function checkPaySystem () {
+      switch(numberField.value[0]) {
+        case '2': cardIcon.src = 'https://photo.virtualbrest.ru/uploads/2022/03/15/c29822b85d1ba85616b98c231c73119c.jpeg';
+        break;
+        case '3': cardIcon.src = 'https://pbs.twimg.com/profile_images/983285404253196288/rx3n00Ep_400x400.jpg';
+        break;
+        case '4': cardIcon.src = 'https://infocity.tech/wp-content/uploads/2020/07/Visa-logo.jpg';
+        break;
+        case '5': cardIcon.src = 'https://models.rsbis.com/storage/makets/preview/resize_600x600/26/42/2642b70ee46621ca320c3a82fc9fdc71.jpeg';
+        break;
+        default: cardIcon.src = 'https://play-lh.googleusercontent.com/baXy546Srucl3vM1yaHr060eBL9_mrk0NH2GGRRCMTrKbekbx2pI77WCaXmNwUqnqQ';
+        break;
+      }      
+    }
 
     function checkCardNumber() {
       const maxLength = 16;
       if (numberField.value.length > maxLength) {
         numberField.value = numberField.value.slice(0, maxLength);
-      }
-      if(numberField.value[0] === '4') {
-        console.log('da')
-        alert('visa');
-      }
-      
+      }    
+      checkPaySystem();    
     }
     numberField.addEventListener('input', checkCardNumber);
 
@@ -135,26 +263,35 @@ class orderWindow {
       const maxLength = 3;
       if (cvvInput.value.length > maxLength) {
         cvvInput.value = cvvInput.value.slice(0, maxLength);
-      }
-      if (cvvInput.value.length === maxLength)  cvvInput.style.borderColor = '#46760A';
-      else cvvInput.style.borderColor = 'red';
-      
+      }      
     }
     cvvInput.addEventListener('input', checkCVV);
 
     //date
     function dateCheck() {
       const maxLength = 5;
-      if (dateInput.value.length === 2) {
-        dateInput.value += '/';
-      }
+      let cardDate = dateInput.value.replace(/\D/g, '').substring(0,4);
+      const matchValue = cardDate.match(/.{1,2}/g);
+      if (matchValue) {
+        cardDate = matchValue.join('/');
+      }       
+      dateInput.value = cardDate;
       if (dateInput.value.length > maxLength) {
         dateInput.value = dateInput.value.slice(0, maxLength);
       }
-      if (dateInput.value.length === maxLength)  dateInput.style.borderColor = '#46760A';
-      else dateInput.style.borderColor = 'red';
+   
     }
     dateInput.addEventListener('input', dateCheck);
+
+    
+    const closeBack = <HTMLElement>document.querySelector('.modal-back-open');
+    const closeWindow = <HTMLElement>document.querySelector('.modal');
+
+    closeBack.addEventListener('click', function () {
+      closeWindow.remove();
+      closeBack.remove();
+    })
+    
   }
 }
