+++ src/style/components/_order.scss
@@ -1,8 +1,25 @@
 .modal {
-  max-width: 500px;
-  border: 2px solid #333333;
-  border-radius: 10px;
-  margin: 0 auto;
+  // max-width: 500px;
+  // border: 2px solid #333333;
+  // border-radius: 10px;
+  // margin: 0 auto;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  -webkit-transform: translate(-50%, -50%);
+  transform: translate(-50%, -50%);
+  width: 500px;
+  //height: 80vh;
+  background-color: #F5F7F6;
+  margin: auto auto;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+  -ms-flex-direction: column;
+  flex-direction: column;
+  z-index: 3;
 }
 
 .modal__title,
@@ -99,4 +116,19 @@
   color: rgb(203, 21, 21);
   width: 80%;
   margin: 1% auto 0;
+}
+
+.modal-back {
+  display: none;
+}
+
+.modal-back-open {
+  display: block;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(41, 41, 41, 0.6);
+  z-index: 2;
 }
